issue #119 refactor: UseCase 를 :core:domain 으로 이동, @Inject 생성자 전환
domain/ 의 UseCase 8개를 :core:domain 으로 이동하고 @Inject 생성자로
전환한다. Repository 인터페이스가 :core:data 에 바인딩되어 Hilt 가 자동
제공하므로, app 잔류 5개 모듈(Album/Friend/Login/Photo/SubAlbum)의
UseCase provider 를 제거했다(중복 바인딩 방지). ViewModel 의 UseCase
import 를 core.domain 으로 갱신했다.
