+++ app/src/main/java/cord/eoeo/momentwo/di/FriendModule.kt
@@ -1,6 +1,6 @@
 package cord.eoeo.momentwo.di
-import cord.eoeo.momentwo.core.data.di.QualifierModule
 
+import cord.eoeo.momentwo.core.data.di.QualifierModule
 import cord.eoeo.momentwo.core.database.MomentwoDatabase
 import cord.eoeo.momentwo.core.data.friend.FriendDataSource
 import cord.eoeo.momentwo.core.data.friend.FriendRemoteMediator
@@ -11,7 +11,6 @@ import cord.eoeo.momentwo.core.database.FriendRemoteKeyDao
 import cord.eoeo.momentwo.core.data.friend.remote.FriendRemoteDataSource
 import cord.eoeo.momentwo.core.data.friend.remote.FriendService
 import cord.eoeo.momentwo.core.data.friend.FriendRepository
-import cord.eoeo.momentwo.domain.friend.GetFriendListUseCase
 import dagger.Module
 import dagger.Provides
 import dagger.hilt.InstallIn
@@ -65,9 +64,4 @@ object FriendModule {
         friendRemoteMediator: FriendRemoteMediator,
     ): FriendRepository =
         FriendRepositoryImpl(friendRemoteDataSource, friendLocalDataSource, friendRemoteMediator)
-
-    @Provides
-    @Singleton
-    fun provideGetFriendListUseCase(friendRepository: FriendRepository) =
-        GetFriendListUseCase(friendRepository)
 }
