+++ app/src/main/java/cord/eoeo/momentwo/di/LoginModule.kt
@@ -1,15 +1,11 @@
 package cord.eoeo.momentwo.di
 
-import cord.eoeo.momentwo.core.datastore.PreferenceRepository
 import cord.eoeo.momentwo.core.data.login.LoginDataSource
 import cord.eoeo.momentwo.core.data.login.LoginRepositoryImpl
 import cord.eoeo.momentwo.core.data.login.remote.LoginRemoteDataSource
 import cord.eoeo.momentwo.core.data.login.remote.LoginService
 import cord.eoeo.momentwo.core.data.login.LoginRepository
-import cord.eoeo.momentwo.domain.login.RequestLoginUseCase
-import cord.eoeo.momentwo.domain.login.TryAutoLoginUseCase
 import cord.eoeo.momentwo.core.data.mapper.ProfileMapper
-import cord.eoeo.momentwo.core.data.profile.ProfileRepository
 import dagger.Module
 import dagger.Provides
 import dagger.hilt.InstallIn
@@ -34,17 +30,4 @@ object LoginModule {
         loginRemoteDataSource: LoginDataSource,
         profileMapper: ProfileMapper,
     ): LoginRepository = LoginRepositoryImpl(loginRemoteDataSource, profileMapper)
-
-    @Provides
-    @Singleton
-    fun provideRequestLoginUseCase(
-        loginRepository: LoginRepository,
-        profileRepository: ProfileRepository,
-        preferenceRepository: PreferenceRepository,
-    ): RequestLoginUseCase = RequestLoginUseCase(loginRepository, profileRepository, preferenceRepository)
-
-    @Provides
-    @Singleton
-    fun provideTryAutoLoginUseCase(preferenceRepository: PreferenceRepository): TryAutoLoginUseCase =
-        TryAutoLoginUseCase(preferenceRepository)
 }
