+++ app/src/main/java/cord/eoeo/momentwo/di/SubAlbumModule.kt
@@ -4,7 +4,6 @@ import cord.eoeo.momentwo.core.data.subalbum.SubAlbumDataSource
 import cord.eoeo.momentwo.core.data.subalbum.SubAlbumRepositoryImpl
 import cord.eoeo.momentwo.core.data.subalbum.remote.SubAlbumRemoteDataSource
 import cord.eoeo.momentwo.core.data.subalbum.remote.SubAlbumService
-import cord.eoeo.momentwo.domain.subalbum.ChangeSubAlbumTitleUseCase
 import cord.eoeo.momentwo.core.data.subalbum.SubAlbumRepository
 import dagger.Module
 import dagger.Provides
@@ -29,9 +28,4 @@ object SubAlbumModule {
     @Singleton
     fun provideSubAlbumRepository(subAlbumRemoteDataSource: SubAlbumDataSource): SubAlbumRepository =
         SubAlbumRepositoryImpl(subAlbumRemoteDataSource)
-
-    @Provides
-    @Singleton
-    fun provideChangeSubAlbumTitleUseCase(subAlbumRepository: SubAlbumRepository): ChangeSubAlbumTitleUseCase =
-        ChangeSubAlbumTitleUseCase(subAlbumRepository)
 }
