+++ app/src/main/java/cord/eoeo/momentwo/ui/photolist/PhotoListViewModel.kt
@@ -5,7 +5,7 @@ import androidx.lifecycle.SavedStateHandle
 import androidx.lifecycle.viewModelScope
 import androidx.navigation.toRoute
 import cord.eoeo.momentwo.core.data.photo.PhotoRepository
-import cord.eoeo.momentwo.domain.subalbum.ChangeSubAlbumTitleUseCase
+import cord.eoeo.momentwo.core.domain.subalbum.ChangeSubAlbumTitleUseCase
 import cord.eoeo.momentwo.core.common.BaseViewModel
 import cord.eoeo.momentwo.ui.MomentwoDestination
 import cord.eoeo.momentwo.ui.photolist.PhotoListContract.Effect.ShowSnackbar
