+++ core/domain/src/main/java/cord/eoeo/momentwo/core/domain/album/RequestCreateAlbumUseCase.kt
@@ -1,7 +1,8 @@
-package cord.eoeo.momentwo.domain.album
+package cord.eoeo.momentwo.core.domain.album
+import javax.inject.Inject
 import cord.eoeo.momentwo.core.data.album.AlbumRepository
 
-class RequestCreateAlbumUseCase(
+class RequestCreateAlbumUseCase @Inject constructor(
     private val albumRepository: AlbumRepository,
 ) {
     suspend operator fun invoke(
