+++ core/domain/src/main/java/cord/eoeo/momentwo/core/domain/friend/GetFriendListUseCase.kt
@@ -1,9 +1,10 @@
-package cord.eoeo.momentwo.domain.friend
+package cord.eoeo.momentwo.core.domain.friend
+import javax.inject.Inject
 import cord.eoeo.momentwo.core.data.friend.FriendRepository
 
 import cord.eoeo.momentwo.core.model.FriendItem
 
-class GetFriendListUseCase(
+class GetFriendListUseCase @Inject constructor(
     private val friendRepository: FriendRepository,
 ) {
     suspend operator fun invoke(): Result<List<FriendItem>> = friendRepository.getFriendList()
