+++ core/domain/src/main/java/cord/eoeo/momentwo/core/domain/login/RequestLoginUseCase.kt
@@ -1,10 +1,11 @@
-package cord.eoeo.momentwo.domain.login
+package cord.eoeo.momentwo.core.domain.login
+import javax.inject.Inject
 import cord.eoeo.momentwo.core.data.login.LoginRepository
 
 import cord.eoeo.momentwo.core.datastore.PreferenceRepository
 import cord.eoeo.momentwo.core.data.profile.ProfileRepository
 
-class RequestLoginUseCase(
+class RequestLoginUseCase @Inject constructor(
     private val loginRepository: LoginRepository,
     private val profileRepository: ProfileRepository,
     private val preferenceRepository: PreferenceRepository,
