+++ core/domain/src/main/java/cord/eoeo/momentwo/core/domain/login/TryAutoLoginUseCase.kt
@@ -1,9 +1,10 @@
-package cord.eoeo.momentwo.domain.login
+package cord.eoeo.momentwo.core.domain.login
+import javax.inject.Inject
 import cord.eoeo.momentwo.core.data.login.LoginRepository
 
 import cord.eoeo.momentwo.core.datastore.PreferenceRepository
 
-class TryAutoLoginUseCase(
+class TryAutoLoginUseCase @Inject constructor(
     private val preferenceRepository: PreferenceRepository,
 ) {
     suspend operator fun invoke(): Result<Unit> =
