+++ core/domain/src/main/java/cord/eoeo/momentwo/core/domain/photo/DownloadPhotoUseCase.kt
@@ -1,7 +1,8 @@
-package cord.eoeo.momentwo.domain.photo
+package cord.eoeo.momentwo.core.domain.photo
+import javax.inject.Inject
 import cord.eoeo.momentwo.core.data.photo.PhotoRepository
 
-class DownloadPhotoUseCase(private val photoRepository: PhotoRepository) {
+class DownloadPhotoUseCase @Inject constructor(private val photoRepository: PhotoRepository) {
     suspend operator fun invoke(imageUrl: String): Result<Unit> =
         photoRepository.downloadPhoto(imageUrl)
 }
