+++ core/domain/src/main/java/cord/eoeo/momentwo/core/domain/photo/UpdateIsLikedUseCase.kt
@@ -1,7 +1,8 @@
-package cord.eoeo.momentwo.domain.photo
+package cord.eoeo.momentwo.core.domain.photo
+import javax.inject.Inject
 import cord.eoeo.momentwo.core.data.photo.PhotoRepository
 
-class UpdateIsLikedUseCase(private val photoRepository: PhotoRepository) {
+class UpdateIsLikedUseCase @Inject constructor(private val photoRepository: PhotoRepository) {
     suspend operator fun invoke(photoId: Int, isLiked: Boolean) =
         photoRepository.updateIsLiked(photoId, isLiked)
 }
